+++ frontend/src/components/UserGenPage.js
@@ -245,7 +245,7 @@ class UserGenPage extends Component {
                     width: `${201 * fontSizeFactor}px`,
                   }}
                   tooltip={t('This is your trading avatar')}
-                  tooltipPosition='up'
+                  tooltipPosition='top'
                 />
                 <br />
               </Grid>
